backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docutils==0.21.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mailersend==0.6.0
packaging==25.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
